requirements/base.txt
beautifulsoup4==4.12.3
celery==5.3.6
Django==5.0.1
django-anymail==10.2
django-allauth==0.61.1
django-celery-results==2.5.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
ipython==8.20.0
google-auth-oauthlib==1.2.0
redis==5.0.1
pillow==10.2.0
